+++ lib/obd/native/fakeTiming.h
@@ -16,13 +16,13 @@ void initTimer();
  * @brief Get the amount milliseconds since start of program
  * @return The milliseconds since start of program
  */
-unsigned long millis();
+uint32_t millis();
 
 /**
  * @brief Get the amount microseconds since start of program
  * @return The microseconds since start of program
  */
-unsigned long micros();
+uint32_t micros();
 
 /**
  * @brief Get the amount microseconds since start of program
@@ -34,7 +34,7 @@ uint64_t micros64();
  * @brief Wait before next execution
  * @param ms Amount of millisecond to wait
  */
-void delay(unsigned long);
+void delay(uint32_t);
 
 /**
  * @brief Wait before next execution
